10_october/10_04_25/requirements.txt
fastapi 
uvicorn 
python-multipart 
openai-whisper 
python-dotenv